requirements.txt
click==8.1.3
cx-Oracle==8.3.0
greenlet==2.0.2
python-dotenv==1.0.0
services==0.1.1
SQLAlchemy==1.4.48
typing_extensions==4.5.0
